consolidate util.go and pubkey.go to validator.go
